backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
asyncpg==0.29.0
sqlalchemy>=1.4.42,<3.0
databases[postgresql]>=0.8.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Document Processing
pypdf==3.17.2
python-docx==1.1.0
python-pptx==0.6.23

# NLP and AI
spacy==3.7.2
nltk==3.8.1
transformers==4.36.2
torch==2.2.0
networkx==3.2.1
scikit-learn==1.3.2

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Configuration & Environment
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
loguru==0.7.2
redis==5.0.1
celery==5.3.4

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
